\name{gazelleRelocations}
\docType{data}
\alias{gazelleRelocations}
\title{Relocations of Mongolian Gazelles}
\description{
    Data set of GPS tracked relocations of gazelles in the Mongolian
    steppe.  For more information contact Thomas Mueller (e-mail 
    address see below). 
}
\usage{gazelleRelocations}
\format{
    Data frame with columns \emph{id}, \emph{sex}, \emph{time}, \emph{x} and \emph{y}.
}
\source{Mueller, T. (2015), muellert[AT]gmail.com}
\keyword{datasets}
