\name{SyncMove-package}
\docType{package}
\alias{SyncMove-package}
\alias{SyncMove}
\title{
    Subsample Temporal Data to Synchronal Events and Compute the MCI
}
\description{
    The function \code{syncSubsample} subsamples temporal data of different
    entities so that the result only contains  synchronal events. The
    function \code{mci} calculates the Movement Coordination Index (MCI, see
                                                                    reference on help page for function \code{mci}) e.g. of a data set
    created with the function \code{syncSubsample}.
}
\details{
    \tabular{ll}{
        Package: \tab SyncMove\cr
        Type: \tab Package\cr
        Version: \tab 0.1-0\cr
        Date: \tab 2015-10-08\cr
        License: \tab GLP-2\cr
    }
}
\author{
    Martin Rimmler (maintainer, martin.rimmler[AT]gmail.com),
    Thomas Mueller
}
\references{
    Mueller, T., et al. (2011) How landscape dynamics link
    individual- to population-level movement patterns: 
        a multispecies comparison of ungulate relocation data,
    Global Ecology and Biogeography, \bold{20}, pages 683--694.
}